+++ app/src/main/res/layout/activity_edit_time_table_two.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".timetable.EditTimeTableTwoActivity">
+    tools:context=".timetable.edit.EditTimeTableTwoActivity">
 
     <TextView
         android:id="@+id/tv_title"
